requirements.txt
asgiref==3.2.7
astroid==2.4.1
boto3==1.14.1
botocore==1.17.1
Django==3.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.3.0
django-filter==2.3.0
django-ses==1.0.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
drf-nested-routers==0.91
future==0.18.2
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
wrapt==1.12.1
